moved the setup of the name column to its own class
